Seperated DAOs into their respectice models
