+++ eventdao.py
@@ -1,0 +1,14 @@
+from dao import DAO
+
+class EventDAO(DAO):
+    def get(self, event_name):
+        return #self._get_by_params('event', {'name': event_name})
+
+    def create(self, event):
+        return
+
+    def update(self, event):
+        return
+
+    def delete(self, event_name):
+        return
