+++ scheduledao.py
@@ -1,0 +1,8 @@
+from dao import DAO
+
+class ScheduleDAO(DAO):
+    def get(self, owner):
+        return
+
+    def update(self, schedule):
+        return
