+++ time.py
@@ -1,0 +1,17 @@
+from typing import Dict
+
+class Time:
+    def __init__(self, hour: int, minute: int):
+        assert 0 <= minute < 60 and 0 <= hour < 24
+        self.hour = hour
+        self.minute = minute
+
+    @classmethod
+    def fromdict(cls, time_dict: Dict[str, int]):
+        return cls(time_dict['hour'], time_dict['minute'])
+
+    def in_seconds(self):
+        return self.hour*60 + self.minute
+
+    def __str__(self):
+        return '%02d:%02d' % (self.hour, self.minute)
